improve FineModelChangeListener
