+++ src/main/java/spoon/compiler/Environment.java
@@ -332,5 +332,13 @@ public interface Environment {
 	/** Set the directory where binary .class files are created */
 	String getBinaryOutputDirectory();
 
+	/**
+	 * get the model change listener that is used to follow the change of the AST.
+	 */
 	FineModelChangeListener getModelChangeListener();
+
+	/**
+	 * set the model change listener
+	 */
+	void setModelChangeListener(FineModelChangeListener modelChangeListener);
 }
