+++ src/main/java/spoon/support/StandardEnvironment.java
@@ -489,6 +489,7 @@ public class StandardEnvironment implements Serializable, Environment {
 		return modelChangeListener;
 	}
 
+	@Override
 	public void setModelChangeListener(FineModelChangeListener modelChangeListener) {
 		this.modelChangeListener = modelChangeListener;
 	}
